requirements-dev.txt
flake8
pytest
coverage
tox
isort
pyramid_debugtoolbar
sqlalchemy
